$6072 move shipping classes in the same package
https://github.com/metasfresh/metasfresh/issues/6072
