+++ de.metas.business/src/main/java/de/metas/shipping/model/MMShipperTransportation.java
