+++ de.metas.business/src/main/java/de/metas/shipping/model/MMShippingPackage.java
