+++ de.metas.business/src/main/java/de/metas/shipping/package-info.java
@@ -1,0 +1,1 @@
+package de.metas.shipping;
