+++ de.metas.business/src/main/java/de/metas/shipping/util/Constants.java
